Modification de inrange pour qu'elle accepte un entier en paramètre

+++ Carpooling-Spring-WS/src/main/java/fr/ups/carpooling/domain/User.java
@@ -113,7 +113,7 @@ public class User {
         return (rad * 180 / Math.PI);
     }
 
-    public Boolean inrange(User voisin, Double range) {
+    public Boolean inrange(User voisin, Integer range) {
         Double dist = this.distance(Double.valueOf(voisin.getLatitude()),
                 Double.valueOf(voisin.getLongitude()));
         System.out.println(dist);
